backend/apps/users/api/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers


class AgentSignupSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True)


class AgentOTPVerifySerializer(serializers.Serializer):
    email = serializers.EmailField()
    otp = serializers.CharField()

