backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
uvicorn==0.27.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
redis==5.0.1


